version 0.0.4

+++ Rakefile
@@ -54,11 +54,18 @@ task :post do
     puts "Error - date format must be YYYY-MM-DD, please check you typed it correctly!"
     exit -1
   end
+  # 新增，首先判断分类目录是否存在，不存在则创建
+  filename = File.join(CONFIG['posts'], category)
+  if !File.directory?(filename)
+    mkdir_p filename
+  end
   filename = File.join(CONFIG['posts'], "#{date}-#{slug}.#{CONFIG['post_ext']}")
   if File.exist?(filename)
     abort("rake aborted!") if ask("#{filename} already exists. Do you want to overwrite?", ['y', 'n']) == 'n'
   end
-  
+  # 新增用户提示，在创建博客之前最后再检查一次是否按照自己的需求正确创建
+  # User confirm 
+  abort("rake aborted!") if ask("The post #{filename} will be created in category #{category}, are you sure?", ['y', 'n']) == 'n'
   puts "Creating new post: #{filename}"
   open(filename, 'w') do |post|
     post.puts "---"
